+++ automations/js/package.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "@octokit/rest": "18.12.0",
     "js-yaml": "^4.1.0",
+    "k6": "0.0.0",
     "node-fetch": "^3.0.0"
   },
   "volta": {
